fetchを分離
